+++ package.json
@@ -23,6 +23,8 @@
   "dependencies": {
     "@dsnp/parquetjs": "^1.1.2",
     "@elastic/elasticsearch": "^7.15.0",
+    "@types/lodash.debounce": "^4.0.6",
+    "@types/react-syntax-highlighter": "^13.5.2",
     "ace-builds": "^1.4.13",
     "alasql": "^1.7.3",
     "chart.js": "^3.5.1",
@@ -66,27 +68,18 @@
   "devDependencies": {
     "@sucrase/jest-plugin": "^2.1.1",
     "@types/cookie-parser": "^1.4.2",
-    "@types/electron": "^1.6.10",
     "@types/express": "^4.17.13",
     "@types/express-session": "^1.17.4",
     "@types/jsesc": "^3.0.1",
     "@types/json-stringify-safe": "^5.0.0",
-    "@types/lodash.debounce": "^4.0.6",
     "@types/mssql": "^7.1.3",
-    "@types/node-fetch": "^3.0.3",
-    "@types/openid-client": "^3.7.0",
-    "@types/oracledb": "^5.2.1",
-    "@types/papaparse": "^5.2.6",
+    "@types/papaparse": "^5.3.0",
     "@types/passport": "^1.0.7",
     "@types/pg": "^8.6.1",
-    "@types/react": "^17.0.27",
+    "@types/react": "^17.0.30",
     "@types/react-dom": "^17.0.9",
-    "@types/react-syntax-highlighter": "^13.5.2",
-    "@types/refractor": "^3.0.2",
-    "@types/snowflake-sdk": "^1.6.1",
     "@types/ssh2-sftp-client": "^7.0.0",
     "@types/uuid": "^8.3.1",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.3",
     "electron": "^15.1.0",
     "electron-chromedriver": "^15.0.0",
     "electron-packager": "^15.4.0",
