+++ ui/Dashboard.tsx
@@ -1,0 +1,1 @@
+export function Dashboard() {}
