+++ ui/NotFound.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export function NotFound() {
+  return <p>Page not found!</p>;
+}
