+++ ui/app.tsx
@@ -1,41 +1,14 @@
-import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 import * as React from 'react';
-import {
-  APP_NAME,
-  MODE,
-  MODE_FEATURES,
-  SITE_ROOT,
-  VERSION,
-} from '../shared/constants';
+import { MODE, MODE_FEATURES, VERSION } from '../shared/constants';
 import { LANGUAGES } from '../shared/languages';
 import log from '../shared/log';
 import '../shared/polyfill';
-import {
-  GetProjectsRequest,
-  GetProjectsResponse,
-  MakeProjectRequest,
-  MakeProjectResponse,
-  OpenProjectRequest,
-  OpenProjectResponse,
-} from '../shared/rpc';
-import {
-  ConnectorInfo,
-  DEFAULT_PROJECT,
-  ProjectPage,
-  ProjectState,
-  ServerInfo,
-} from '../shared/state';
-import { asyncRPC } from './asyncRPC';
-import { Alert } from './components/Alert';
-import { Button } from './components/Button';
-import { ErrorBoundary } from './components/ErrorBoundary';
-import { Input } from './components/Input';
-import { ConnectorList } from './ConnectorList';
-import { PageList } from './PageList';
+import { DEFAULT_PROJECT, ProjectState } from '../shared/state';
+import { Dashboard } from './Dashboard';
+import { Editor } from './Editor';
+import { NotFound } from './NotFound';
 import { makeStore, ProjectContext, ProjectStore } from './ProjectStore';
-import { ServerList } from './ServerList';
-import { Sidebar } from './Sidebar';
-import { Updates } from './Updates';
+import { UrlStateContext, useUrlState } from './urlState';
 
 if (MODE === 'browser') {
   Object.values(LANGUAGES).map((l) => {
@@ -43,20 +16,6 @@ if (MODE === 'browser') {
       l.inMemoryInit();
     }
   });
-}
-
-function getQueryParameter(param: String) {
-  const query = window.location.search.substring(1);
-  const vars = query.split('&');
-
-  for (let i = 0; i < vars.length; i++) {
-    const pair = vars[i].split('=');
-    if (pair[0] === param) {
-      return decodeURIComponent(pair[1]);
-    }
-  }
-
-  return '';
 }
 
 function useProjectState(
@@ -131,54 +90,8 @@ function useProjectState(
 const store = makeStore(MODE);
 
 export function App() {
-  const requestedProjectId = getQueryParameter('project');
-  const [projectId, setProjectId] = React.useState(
-    requestedProjectId ||
-      (MODE_FEATURES.useDefaultProject ? DEFAULT_PROJECT.projectName : '')
-  );
-  (window as any).projectId = projectId;
-  React.useEffect(() => {
-    if (!requestedProjectId && projectId && MODE !== 'browser') {
-      window.location.href = window.location.pathname + '?project=' + projectId;
-    }
-  }, [requestedProjectId, projectId]);
-
-  const [makeProjectError, setMakeProjectError] = React.useState('');
-  async function makeProject(projectId: string) {
-    try {
-      await asyncRPC<MakeProjectRequest, MakeProjectResponse>('makeProject', {
-        projectId,
-      });
-      setProjectId(projectId);
-    } catch (e) {
-      setMakeProjectError(e.message);
-    }
-  }
-
-  const [state, updateProjectState] = useProjectState(projectId, store);
-
-  const currentPageKey = 'currentPage:' + projectId;
-  const [currentPage, _setCurrentPage] = React.useState(
-    +localStorage.getItem(currentPageKey) || 0
-  );
-  function setCurrentPage(p: number) {
-    localStorage.setItem(currentPageKey, String(p) || '0');
-    return _setCurrentPage(p);
-  }
-
-  React.useEffect(() => {
-    if (state && state.projectName) {
-      log.info(state);
-      document.title = state.projectName;
-    }
-
-    // Set body overflow once on init
-    if (MODE_FEATURES.noBodyYOverflow) {
-      document.body.style.overflowY = 'hidden';
-    }
-  }, [state && state.projectName]);
-
-  const [projectNameTmp, setProjectNameTmp] = React.useState('');
+  const [urlState, setUrlState] = useUrlState();
+  const [state, setProjectState] = useProjectState(projectId, store);
 
   const [headerHeight, setHeaderHeightInternal] = React.useState(0);
   const setHeaderHeight = React.useCallback((e: HTMLElement) => {
@@ -189,247 +102,29 @@ export function App() {
     setHeaderHeightInternal(e.offsetHeight);
   }, []);
 
-  const [projects, setProjects] = React.useState<GetProjectsResponse | null>(
-    null
-  );
-  React.useEffect(() => {
-    async function load() {
-      const projects = await asyncRPC<GetProjectsRequest, GetProjectsResponse>(
-        'getProjects',
-        null
-      );
-      setProjects(projects);
-    }
-
-    if (MODE === 'server' && !projects) {
-      load();
-    }
-  }, []);
-
-  if ((!state && projectId) || (MODE === 'server' && !projects)) {
-    return (
-      <div className="loading">
-        Loading...
-        <span id="spin"></span>
-      </div>
-    );
-  }
-
-  function updatePage(page: ProjectPage) {
-    state.pages[currentPage] = page;
-    updateProjectState(state);
-  }
-
-  function addPage(page: ProjectPage) {
-    state.pages.push(page);
-    updateProjectState(state);
-  }
-
-  function deletePage(at: number) {
-    state.pages.splice(at, 1);
-    updateProjectState(state);
-  }
-
-  function updateConnector(id: string, dc: ConnectorInfo) {
-    const index = (state.connectors || []).findIndex((c) => c.id === id);
-    if (index === -1) {
-      state.connectors.push(dc);
-      return;
-    }
-    state.connectors[index] = dc;
-    updateProjectState(state);
-  }
-
-  function addConnector(dc: ConnectorInfo) {
-    state.connectors.push(dc);
-    updateProjectState(state);
-  }
-
-  function deleteConnector(id: string) {
-    const at = (state.connectors || []).findIndex((c) => c.id === id);
-    if (at === -1) {
-      return;
-    }
-    state.connectors.splice(at, 1);
-    updateProjectState(state);
-  }
-
-  function updateServer(id: string, dc: ServerInfo) {
-    const index = (state.servers || []).findIndex((c) => c.id === id);
-    if (index === -1) {
-      state.servers.push(dc);
-      return;
-    }
-    state.servers[index] = dc;
-    updateProjectState(state);
-  }
-
-  function addServer(dc: ServerInfo) {
-    state.servers.push(dc);
-    updateProjectState(state);
-  }
-
-  function deleteServer(id: string) {
-    const at = (state.servers || []).findIndex((c) => c.id === id);
-    if (at === -1) {
-      return;
-    }
-    state.servers.splice(at, 1);
-    updateProjectState(state);
-  }
-
-  async function openProject() {
-    await asyncRPC<OpenProjectRequest, OpenProjectResponse>(
-      'openProject',
-      null
-    );
-    window.close();
-  }
-
-  // This allows us to render the sidebar in tests where we
-  // prepopulate connectors and servers
-  const hasSidebar = Boolean(
-    MODE_FEATURES.connectors ||
-      state.connectors?.length ||
-      state.servers?.length
-  );
+  const body =
+    {
+      editor: Editor,
+      dashboard: Dashboard,
+    }[urlState.view || 'editor'] || NotFound;
 
   return (
-    <ProjectContext.Provider value={state}>
-      <div className={`app app--${MODE}`}>
-        {MODE_FEATURES.appHeader && (
-          <header ref={setHeaderHeight}>
-            <div className="vertical-align-center">
-              <span className="logo">{APP_NAME}</span>
-              <div className="flex-right vertical-align-center">
-                {MODE === 'browser' ? (
-                  <React.Fragment>
-                    <span title="Drop all state and load a sample project.">
-                      <Button
-                        onClick={() => {
-                          updateProjectState(DEFAULT_PROJECT);
-                          window.location.reload();
-                        }}
-                      >
-                        Reset
-                      </Button>
-                    </span>
-                    <a
-                      href="https://github.com/multiprocessio/datastation"
-                      target="_blank"
-                    >
-                      <iframe
-                        src="https://ghbtns.com/github-btn.html?user=multiprocessio&repo=datastation&type=star&count=true&size=medium"
-                        frameBorder="0"
-                        scrolling="0"
-                        width="80"
-                        height="20"
-                        title="GitHub"
-                      ></iframe>
-                    </a>
-                    <a
-                      href={`${SITE_ROOT}/#online-environment`}
-                      target="_blank"
-                    >
-                      About
-                    </a>
-                  </React.Fragment>
-                ) : (
-                  <span>{projectId}</span>
-                )}
-              </div>
-            </div>
-          </header>
-        )}
-        <main
-          style={{
-            marginTop: headerHeight,
-            height: `calc(100% - ${headerHeight}px)`,
-          }}
-        >
-          {projectId && hasSidebar && (
-            <Sidebar>
-              <ConnectorList
-                state={state}
-                updateConnector={updateConnector}
-                addConnector={addConnector}
-                deleteConnector={deleteConnector}
-              />
-              <ServerList
-                state={state}
-                updateServer={updateServer}
-                addServer={addServer}
-                deleteServer={deleteServer}
-              />
-              <ErrorBoundary>
-                <Updates />
-              </ErrorBoundary>
-            </Sidebar>
+    <ProjectContext.Provider value={{ state, setProjectState }}>
+      <UrlStateContext.Provider value={{ urlState, setUrlState }}>
+        <div className={`app app--${MODE}`}>
+          {MODE_FEATURES.appHeader && (
+            <Header setHeaderHeight={setHeaderHeight} />
           )}
-          <div className="main-body">
-            {!projectId ? (
-              <div className="project-name">
-                <h1>New Project</h1>
-                <p>Pick a name for this project to get started.</p>
-                <div className="form-row">
-                  <Input
-                    value={projectNameTmp}
-                    label="Project name"
-                    onChange={(v) => setProjectNameTmp(v)}
-                  />
-                </div>
-                <div className="form-row">
-                  <Button
-                    type="primary"
-                    disabled={!projectNameTmp}
-                    onClick={() => makeProject(projectNameTmp)}
-                  >
-                    {projectNameTmp ? 'Go!' : 'Pick a name'}
-                  </Button>
-                </div>
-                {makeProjectError && (
-                  <Alert type="error" children={makeProjectError} />
-                )}
-                {MODE === 'desktop' && (
-                  <div className="project-existing">
-                    <p>Or open an existing project.</p>
-                    <div className="form-row">
-                      <Button onClick={openProject}>Open</Button>
-                    </div>
-                  </div>
-                )}
-                {MODE === 'server' && projects.length ? (
-                  <div className="project-existing">
-                    <p>Or open an existing project.</p>
-                    {projects.map(({ name, createdAt }) => (
-                      <div className="form-row">
-                        <h3>{name}</h3>
-                        <div>
-                          Created{' '}
-                          {formatDistanceToNow(new Date(createdAt), {
-                            addSuffix: true,
-                          })}
-                        </div>
-                        <a href={'/?project=' + name}>Open</a>
-                      </div>
-                    ))}
-                  </div>
-                ) : null}
-              </div>
-            ) : (
-              <PageList
-                state={state}
-                updatePage={updatePage}
-                addPage={addPage}
-                deletePage={deletePage}
-                currentPage={currentPage}
-                setCurrentPage={setCurrentPage}
-              />
-            )}
-            <div className="version">{VERSION}</div>
-          </div>
-        </main>
-      </div>
+          <main
+            style={{
+              marginTop: headerHeight,
+              height: `calc(100% - ${headerHeight}px)`,
+            }}
+          >
+            {body}
+          </main>
+        </div>
+      </UrlStateContext.Provider>
     </ProjectContext.Provider>
   );
 }
