+++ ui/components/Version.tsx
@@ -1,0 +1,5 @@
+import { VERSION } from '../../shared/constants';
+
+export function Version() {
+  return <div className="version">{VERSION}</div>;
+}
